refine interfaces methods CxPropertiesBase and CxGoProperties
