+++ src/main/java/com/checkmarx/sdk/service/CxClient.java
@@ -22,7 +22,7 @@ public interface CxClient extends ScannerClient {
     /**
      * Get the last scan Id of a given project Id
      * @param projectId project Id
-     * @return
+     * @return ID of the current scan
      */
     public Integer getLastScanId(Integer projectId);
 
@@ -38,7 +38,7 @@ public interface CxClient extends ScannerClient {
      * Fetches the Timestamp of the last full scan
      *
      * @param projectId
-     * @return
+     * @return Timestamp for when scan started.
      */
     public LocalDateTime getLastScanDate(Integer projectId);
 
@@ -46,7 +46,7 @@ public interface CxClient extends ScannerClient {
      * Get the status of a given scanId
      *
      * @param scanId
-     * @return
+     * @return Status code for current scan ID.
      */
     Integer getScanStatus(Integer scanId);
 
@@ -54,7 +54,7 @@ public interface CxClient extends ScannerClient {
      * Generate a scan report request (xml) based on ScanId
      *
      * @param scanId
-     * @return
+     * @return ID for scan report
      */
     Integer createScanReport(Integer scanId);
 
@@ -62,26 +62,17 @@ public interface CxClient extends ScannerClient {
      * Get the status of a report being generated by reportId
      *
      * @param reportId
-     * @return
+     * @return Status of current report
      */
     Integer getReportStatus(Integer reportId) throws CheckmarxException;
 
-    /**
-     * Retrieve the report by scanId, mapped to ScanResults DTO, applying filtering as requested
-     *
-     * @param scanId
-     * @param filter
-     * @return
-     * @throws CheckmarxException
-     */
-    public ScanResults getReportContentByScanId(Integer scanId, FilterConfiguration filter) throws CheckmarxException;
 
     /**
      * Retrieve the report by reportId, mapped to ScanResults DTO, applying filtering as requested
      *
      * @param reportId
      * @param filter
-     * @return
+     * @return Contents of the current report
      * @throws CheckmarxException
      */
     public ScanResults getReportContent(Integer reportId, FilterConfiguration filter) throws CheckmarxException;
@@ -109,7 +100,7 @@ public interface CxClient extends ScannerClient {
      *
      * @param file
      * @param filter
-     * @return
+     * @return Contents of the current report
      * @throws CheckmarxException
      */
     public ScanResults getReportContent(File file, FilterConfiguration filter) throws CheckmarxException;
@@ -118,7 +109,7 @@ public interface CxClient extends ScannerClient {
      * @param vulnsFile
      * @param libsFile
      * @param filter
-     * @return
+     * @return Contents of current OSA report
      * @throws CheckmarxException
      */
     public ScanResults getOsaReportContent(File vulnsFile, File libsFile, List<Filter> filter) throws CheckmarxException;
@@ -164,14 +155,14 @@ public interface CxClient extends ScannerClient {
     /**
      * Get All Projects in Checkmarx
      *
-     * @return
+     * @return List of projects on server
      */
     public List<CxProject> getProjects() throws CheckmarxException;
 
     /**
      * Get All Projects in Checkmarx
      *
-     * @return
+     * @return List of projects on server
      */
     public List<CxProject> getProjects(String teamId) throws CheckmarxException;
 
@@ -181,7 +172,7 @@ public interface CxClient extends ScannerClient {
      * using TeamId does not work.
      *
      * @param ownerId
-     * @return
+     * @return ID of project
      */
 
     public Integer getProjectId(String ownerId, String name);
@@ -189,7 +180,7 @@ public interface CxClient extends ScannerClient {
     /**
      * Return Project based on projectId
      *
-     * @return
+     * @return Project data
      */
     public CxProject getProject(Integer projectId);
 
@@ -198,7 +189,7 @@ public interface CxClient extends ScannerClient {
      * Check if a scan exists for a projectId
      *
      * @param projectId
-     * @return
+     * @return true if scan exists
      */
     public boolean scanExists(Integer projectId);
 
@@ -216,7 +207,7 @@ public interface CxClient extends ScannerClient {
      * @param projectId
      * @param presetId
      * @param engineConfigId
-     * @return
+     * @return ID for scan settings
      */
     public Integer createScanSetting(Integer projectId, Integer presetId, Integer engineConfigId, Integer postBackId);
 
@@ -224,28 +215,16 @@ public interface CxClient extends ScannerClient {
      * Get Scan Settings for an existing project (JSON String)
      *
      * @param projectId
-     * @return
+     * @return Current scan configuration
      */
     public String getScanSetting(Integer projectId);
 
     CxScanSettings getScanSettingsDto(int projectId);
 
-    /**
-     * Get Preset Name based on Id
-     *
-     * @param presetId
-     * @return
-     */
-    public String getPresetName(Integer presetId);
 
 
-    /**
-     * Get Preset Id of an existing project
-     *
-     * @param projectId
-     * @return
-     */
-    public Integer getProjectPresetId(Integer projectId);
+
+
 
     /**
      * Set Repository details for a project
@@ -295,7 +274,7 @@ public interface CxClient extends ScannerClient {
      * Get teamId for given path
      *
      * @param teamPath
-     * @return
+     * @return ID for team
      * @throws CheckmarxException
      */
     public String getTeamId(String teamPath) throws CheckmarxException;
@@ -338,7 +317,7 @@ public interface CxClient extends ScannerClient {
 
     /**
      * Returns a list of roles in Checkmarx
-     * @return
+     * @return List of roles on server
      */
     public List<CxRole> getRoles() throws CheckmarxException;
 
@@ -351,7 +330,7 @@ public interface CxClient extends ScannerClient {
     /**
      * Retrieve list of Role LDAP mappings associated with an LDAP server Id
      * @param ldapServerId
-     * @return
+     * @return List of ldap roles associated with server
      * @throws CheckmarxException
      */
     public List<CxRoleLdap> getRoleLdap(Integer ldapServerId) throws CheckmarxException;
@@ -360,7 +339,7 @@ public interface CxClient extends ScannerClient {
      * Retrieve the Id of a role mapping associated with an LDAP Group DN
      * @param ldapServerId
      * @param ldapGroupDn
-     * @return
+     * @return Mapping between Cx and LDAP roles
      * @throws CheckmarxException
      */
     public Integer getLdapRoleMapId(Integer ldapServerId, Integer roleId, String ldapGroupDn) throws CheckmarxException;
@@ -421,22 +400,21 @@ public interface CxClient extends ScannerClient {
      */
     public String createTeamWS(String parentTeamId, String teamName) throws CheckmarxException;
 
-    /**
-     * Create team under given parentId - Will use REST API to create team for version 9.0+
-     *
-     * @param parentTeamId
-     * @param teamName
-     * @return new TeamId
-     * @throws CheckmarxException
-     */
-    public String createTeam(String parentTeamId, String teamName) throws CheckmarxException;
+//    /**
+//     * Create team under given parentId - Will use REST API to create team for version 9.0+
+//     *
+//     * @param parentTeamId
+//     * @param teamName
+//     * @return new TeamId
+//     * @throws CheckmarxException
+//     */
+//    public String createTeam(String parentTeamId, String teamName) throws CheckmarxException;
 
     /**
      * Move team to under given newParentId - Will use REST API to create team for version 9.0+
      *
      * @param teamId - Id of the team to be moved
      * @param newParentTeamId - id of the new parent team
-     * @return void
      * @throws CheckmarxException
      */
     public void moveTeam(String teamId, String newParentTeamId) throws CheckmarxException;
@@ -446,7 +424,6 @@ public interface CxClient extends ScannerClient {
      *
      * @param teamId - Id of the team to be renamed
      * @param newTeamName - new team name
-     * @return void
      * @throws CheckmarxException
      */
     public void renameTeam(String teamId, String newTeamName) throws CheckmarxException;
@@ -455,7 +432,6 @@ public interface CxClient extends ScannerClient {
      * Create team under given parentId - Will use REST API to create team for version 9.0+
      *
      * @param teamId
-     * @return
      * @throws CheckmarxException
      */
     public void deleteTeam(String teamId) throws CheckmarxException;
@@ -464,7 +440,7 @@ public interface CxClient extends ScannerClient {
      * Get a team Id based on the name and the Parent Team Id
      * @param parentTeamId
      * @param teamName
-     * @return
+     * @return ID of team
      * @throws CheckmarxException
      */
     public String getTeamId(String parentTeamId, String teamName) throws CheckmarxException;
@@ -474,7 +450,6 @@ public interface CxClient extends ScannerClient {
          * Delete a team based on the name for a given parent team id
          *
          * @param teamId
-         * @return
          * @throws CheckmarxException
          */
     public void deleteTeamWS(String teamId) throws CheckmarxException;
@@ -483,15 +458,15 @@ public interface CxClient extends ScannerClient {
      * Get scan configuration Id by name.
      *
      * @param configuration
-     * @return
+     * @return Scan configuration ID
      * @throws CheckmarxException
      */
     public Integer getScanConfiguration(String configuration) throws CheckmarxException;
 
-    /**
-     * Get scan configuration name by Id.
-     */
-    String getScanConfigurationName(int configurationId);
+//    /**
+//     * Get scan configuration name by Id.
+//     */
+//    String getScanConfigurationName(int configurationId);
 
     /**
      * Fetch the Id of a given preset name
@@ -559,7 +534,6 @@ public interface CxClient extends ScannerClient {
     /**
      *
      * @param scanId
-     * @return
      * @throws CheckmarxException
      */
     public void cancelScan(Integer scanId) throws CheckmarxException;
@@ -600,7 +574,7 @@ public interface CxClient extends ScannerClient {
      * @param teamName
      * @param projectName
      * @param filters
-     * @return
+     * @return Scan results
      * @throws CheckmarxException
      */
     public ScanResults getLatestScanResults(String teamName, String projectName, FilterConfiguration filters) throws CheckmarxException;
