+++ src/main/java/com/cx/restclient/CxGoClientImpl.java
@@ -359,6 +359,7 @@ public class CxGoClientImpl implements ScannerClient {
     }
 
 
+    @Override
     public String createTeam(String parentID, String teamName) throws CheckmarxException {
         return createApplication(teamName, "Generated by CxFlow", parentID);
     }
@@ -1016,5 +1017,26 @@ public class CxGoClientImpl implements ScannerClient {
     @Autowired
     public void setCxRepoFileService(CxRepoFileService cxRepoFileService) {
         this.cxRepoFileService = cxRepoFileService;
+    }
+
+    @Override
+    public Integer getProjectPresetId(Integer projectId) {
+        return UNKNOWN_INT;
+    }
+
+
+    @Override
+    public String getPresetName(Integer presetId) {
+        return null;
+    }
+    
+    @Override
+    public String getScanConfigurationName(int configurationId) {
+        return null;
+    }
+
+    @Override
+    public CxScanSettings getScanSettingsDto(int projectId) {
+        return null;
     }
 }
